requirements.txt
tesseract==0.1.3
pytesseract==0.3.13
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
requests==2.32.5
wheel==0.46.*
Pillow>=10.2.0
numpy==2.0.2
opencv-contrib-python==4.13.0.92
scikit-learn==1.6.1
gdown==5.2.2
pytest-socket==0.7.0
imutils==0.5.4
symspellpy==6.9.0
commitizen==4.10.1
pre-commit==4.3.0
